requirements.txt
cshogi
numpy<2
tqdm
onnxruntime-gpu